@@ -31,7 +31,7 @@
         "source": "https://github.com/picocms/Pico"
     },
     "require": {
-        "php": ">=7.0.8",
+        "php": ">=8.0",
         "ext-mbstring": "*",
         "twig/twig": "^2.12",
         "symfony/yaml" : "^3.4",
